requirements.txt
scipy
numpy
tensorflow-gpu
cvpubsubs